requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==3.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==3.20.1
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2

